多少のリファクタリング
