+++ src/components/templates/Home.tsx
@@ -1,12 +1,11 @@
 import { FC, useState } from 'react';
-import { Box, Container, Heading, Table, TableContainer, Tbody, Td, Tfoot, Th, Thead, Tr } from '@chakra-ui/react';
+import { Box, Container, Heading } from '@chakra-ui/react';
 import { css } from '@emotion/css';
 import type { Todo } from 'data';
 import AddTaskButton from 'components/molecules/AddTaskButton';
 import DeleteEndTaskButton from 'components/molecules/DeleteEndTaskButton';
 import InputTaskName from 'components/molecules/InputTaskName';
-import IsDoneCheckBox from 'components/molecules/IsDoneCheckBox';
-import RemoveTaskButton from 'components/molecules/RemoveTaskButton';
+import TaskTable from 'components/molecules/TaskTable';
 
 type Props = { pageTitle: string };
 
@@ -14,16 +13,10 @@ const infoStyle = css`
 color: #bbb;
 font-size: 12px;
 `
-const isDoneStyle = css`
-text-decoration: line-through;
-color:#bbb
-`
-const doneStyle = (isDone: boolean) => isDone ? isDoneStyle : css``
 
 const h1Style = css`border-bottom: 1px solid #ddd;`
-const ulStyle = css`list-style: none;`
-const liStyle = css``
-
+// const ulStyle = css`list-style: none;`
+// const liStyle = css``
 
 //   const getTextareaStyle = (isValid: boolean) => css`
 //   border: solid 1px ${isValid ? "green" : "red"};
@@ -44,53 +37,21 @@ const Home: FC<Props> = ({ pageTitle }) => {
   //   </li>
   // ))
 
-  const tmp = todos.map((todo, index) => (
-    <Tr key={todo.key}>
-      <Td>{index + 1}</Td>
-      <Td><IsDoneCheckBox setTodos={setTodos} target={todo} /></Td>
-      <Td><span className={doneStyle(todo.isDone)}>{todo.name}</span></Td>
-      <Td><RemoveTaskButton setTodos={setTodos} target={todo} /></Td>
-    </Tr>
-  ))
-
-  const contents = (
-    <TableContainer>
-      <Table size='sm'>
-        <Thead>
-          <Tr>
-            <Th>No.</Th>
-            <Th>isDone</Th>
-            <Th>タスク名</Th>
-            <Th>削除</Th>
-          </Tr>
-        </Thead>
-        <Tbody>
-          {tmp}
-        </Tbody>
-        <Tfoot>
-          <Tr>
-            <Th>isDone</Th>
-            <Th>タスク名</Th>
-            <Th>削除</Th>
-          </Tr>
-        </Tfoot>
-      </Table>
-    </TableContainer>)
-
-
   return (
     <Container py="3" maxW="600px">
       <Heading size="lg" as="h1" my={8}>
         {pageTitle}
       </Heading>
       <h1 className={h1Style}>My Todo Task<span className={infoStyle}>({remaining.length}/{todos.length})</span>
-        <DeleteEndTaskButton setTodos={setTodos} />
+        <DeleteEndTaskButton className={css`float: right;`} setTodos={setTodos} />
       </h1>
       <Box display='flex' mt='2' alignItems='center'>
         <InputTaskName formData={formData} setFormData={setFormData} />
         <AddTaskButton formData={formData} setFormData={setFormData} setTodos={setTodos} />
       </Box>
-      {tmp.length ? contents : <div>Todo なし</div>}
+      {todos.length ?
+        <TaskTable todos={todos} setTodos={setTodos} /> :
+        <div>Todo なし</div>}
     </Container>
   );
 };
